Added info triangle
The site now features a triangle in the upper right corner that opens up an info section modal on click.
The triangle will also display the remaining rate limit, when over 50% was used.

Changed the Modrinth class to adhere to the singleton principle. The class is now accessed via the ModrinthService.