feat(script-generation): add modal in which the user can generate a cleanup & update script
This commits adds a modal in which the user can generate a cleanup & update script for the results of their mods update. The user can choose to generate the script in either powershell or bash flavor.
The script deletes all original files of the updated mods, moves unavailable mods to an "outdated folder" and downloads all available mods.